Use the new media usage index from DokuWiki core if available
Note that you need to rebuild the search index after this change in
order to get correct link updates after media moves, this should happen
automatically for visited pages. The index that is created in the
pagemove plugin is now identical to the one that is created in DokuWiki
core from the Autumn 2013 release on so you won't need to manually
rebuild the index after updating DokuWiki.