rewriting missing links should work now
this introduces stacked locking of the rewrite process which is needed
to keep pages from being rewritten before the missing link info is added
to their metadata.

This works fine for the newly added test case. However in real world
scenarios it might happen that a page is access by someone while the
move operation is in progress. This would still trigger a rewrite,
rendering addition of missing link info too late and create wrong second
rewrites (as seen before this patch in normal operation).

There are multiple ways to fix this:

1) allow absolutely no rewrites during a move (using a file based
   locking instead of process based). This would have the disadvantage
   of exposing pages with wrong links to end users during a move
2) pass the missing link info over to the ops class so that each moved
   page has the info right away and saves it with the other (existing
   link) move data in one go. The problem would then be to apply the
   missing link move data to all pages that were not affected by any
   other link changes. It would need some tracking of already processed
   affected pages vs. all affected pages.

Option 2 is most probably the better way to do it. However this is left
for a later iteration for now as the scope of this project is already
slighly out of bounds.

@michitux I'd be happy about any comments, suggestion or - even better -
code here.