Make helper_plugin_move_plan::stepThroughDocuments unit-testable
The helper_plugin_move_op class-instance was moved to a class object so
we are able to replace it with an mock object in unit-tests. Doing this
we are both able to break the dependency and prevent any pages from
actually being moved and sense the values handed to the
helper_plugin_move_op instance.

The stepThroughDocuments function has now suffcient unit-test coverage
to begin refactoring.