Attach JS rename handler to mobile menu items.
When the MobileMenu is used the required handler to show the rename
dialog wasn't attached to the select resulting in an unkown action
call.

This attaches the handler and uses a jQuery specific trick to be the
first handler to be evaluated (allowing to cancel the default handler
used on the quickselect behaviour).

fixes #150
fixes #169