Fix rewriting of links to the root start page
Previously, we had a bug that the "start"-name was not appended to the id in this case, and we thus didn't detect that the old relative link was still valid. Further, this could have caused issues if the root start page is renamed where certain links wouldn't have been adapted.